ATB 22. delete permission

+++ dmlcomments/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import get_object_or_404, render, redirect
-from django.http import HttpResponseRedirect, HttpResponse
+from django.http import HttpResponseRedirect, HttpResponse, Http404
 from django.urls import reverse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
@@ -8,7 +8,13 @@ from dmlcomments.models import Comment
 from dmlcomments.forms import CommentForm
 
 def comment_thread(request, pk):
-	comment = get_object_or_404(Comment, pk=pk)
+	#comment = get_object_or_404(Comment, pk=pk)
+	try:
+		comment = Comment.objects.get(pk=pk)
+	except:
+		raise Http404
+	if not comment.is_parent:
+		comment = comment.parent # if comment is not a parent, find parent instead
 	content_object = comment.content_object #gets poll/post the comment is from
 	content_id = comment.content_object.id
 	initial_data ={"content_type": comment.content_type,
@@ -39,15 +45,30 @@ def comment_thread(request, pk):
 	context = {'comment': comment, 'form': form}
 	return render(request, 'comment_thread.html', context)
 
-@login_required
+#@login_required #redirects them to login page if not loggedin
 def comment_delete(request, pk):
-	comment = get_object_or_404(Comment, pk=pk)
+	#comment = get_object_or_404(Comment, pk=pk)
+	try:
+		comment = Comment.objects.get(pk=pk)
+	except:
+		raise Http404
+	if comment.author != request.user and request.user.is_staff is False: #check this!
+		#messeges.success(request, "You do not have delete permissions")
+		#return reditect....
+		response = HttpResponse("You do not have delete permissions")
+		response.status_code = 403
+		return response # only let the author delete. Rediirect to "not allowed page"
+
+	#content_object = comment.content_object #gets poll/post the comment is from
+	#content_id = comment.content_object.id
+
 	if request.method == 'POST':
+	  #if not comment.is_parent:  #only delete replies. If parent, replace with "Deleted"
 		parent_obj_url = comment.content_object.get_absolute_url()
 		comment.delete()
 		messages.success(request, "Comment was deleted")
 		return HttpResponseRedirect(parent_obj_url)
-	context = {'comment':comment}
+	context = {'comment':comment,}
 	return render(request, 'confirm_delete.html', context)
 
 @login_required
